Reformat in my own style, and preliminary work for integrating this
No longer includes the 'md5.h' header.  We don't need the HASHSIZE definition
because it's always going to be 16 bytes.

The rest of these changes are intended to limit leakage of names out of
this file into the main 'filter.c' file which will be including it directly.

Removed the macros WORD and MASK.  They were each only used once, so fold
their values in.

The other macros (rotate, F, G, H, I) are all unchanged, but I undefine them
at the bottom of the file.

The array 'T' now has a longer name, but there's a macro defined just in this
file to access it as 'T'.

All the functions now have an 'md5_' prefix, except a few I missed.