Make MD5 implemenation work as an algorithm
This is just for the simple API I have so far, it won't work with streaming
data in in chunks, although I've started heading in that direction.  I've
moved the initialization into an algo init function and stored the context
data in the filter object.

Since my input and output buffers use unsigned chars I've switched to those
throughout the MD5 file to avoid warnings.  That also allows me to remove
some casts.

Tests are included which at least prove that the MD5 implementation is good.