Provide OO interface for loading input data in chunks
There's now a new constructor function in the 'data.filter' table.  Calling
the 'new' method returns a userdata object wrapping a Filter structure,
which can be used to drive one of the algorithms by passing it data with
the 'add' method.  The 'output' method pushes through any unprocessed input
data and then returns the output as a string.

The objects returned by 'new' use a metatable created when the library
is loaded.  I've named it 'data.filter-VERSION', where VERSION is the
version number provided by the makefile (which it gets from the 'Changes'
file).  Probably not much use, but it should help ensure the uniqueness
of the name.

When using a filter this way it needs to use a new output callback called
'output_string', which stores the output in a resizeable output buffer.
The 'output_lbuf' function can't be used because we return to Lua in between
chunks of input, so we can't keep the luaL_Buffer object on the stack.
Because this new output callback resizes the output buffer, the output
functions now need to take a pointer to the end-of-buffer pionter, so that
it can be adjusted if the buffer grows.

The 'output' method sets a new flag in the Filter structure to indicate
that processing has finished.  This acts as a guard against the caller
trying to add more data to an algorithm which has already received a call
with 'eof' set.