Add facility for sending filter output directly to a named file
Now 'data.filter:new()' has an optional argument to specify where the output
should go.  Currently it must be a filename, although I'll add some other
things soon.  Data gets written out with a new output callback, and the file
is automatically closed, with the output buffer having been flushed, when
the object is collected.

I've introduced a new method 'finish' to have the object finish processing
all input, send anything left in the output buffer, and then close the file,
so after calling that the file should contain everything it's going to.
Closing the file will also prevent lots of open file handles sitting around
waiting to be garbage collected.

I've also renamed the 'output' method as 'result', to make it more obvious
that it returns a value.