Fix filehandle leak; test error conditions
I was leaking a file handle when opening the file supplied to 'addfile' was
successful but reading it wasn't.  This happens on Unix systems when you
open a directory for reading.

I've factored out the code for turning binary data into hex text for testing
into the test utils file so that I can use it in multiple test programs.