The data.filter 'addfile' method can now read from file handles
I check the type of the Lua value supplied.  If it's a string then I treat
it like a filename as before.  If it's a table or userdata value, and looking
up the field 'read' returns a function, then I treat it as a file handle
object and repeatedly call the 'read' method to load data from the file and
process it.  The file isn't closed or rewound afterwards, so the only method
it needs to provide is 'read', and that will only be called with a numeric
argument.