Add options for 'base64_encode' algorithm
Both the simple and OO APIs can now accept an optional table of options,
which can control specific aspects of the algorithm's behaviour.

So far there are only options for encoding base64: 'no_padding', if true,
disables the '=' padding characters on the end of the output, and
'line_ending' and 'max_line_length', if either of them is provided, cause
the output to be broken into lines with the given line ending string
attached to the end of each line.

The default maximum line length is 76, and the default line ending is
carriage return line feed, both of which are designed to be appropriate
for email.

This isn't exactly a beautiful implementation.  There's a crufty macro
included after every byte is output to check whether a line ending needs
to be output.