By default, ignore whitespace in 'base64_decode' algorithm
There's now an option 'disallow_whitespace' to enable the old behaviour of
treating a whitespace character as invalid, but the default is to skip over
them, since they often occur as line breaks in large amounts of
base64-encoded data.

When an invalid character is detected, I now throw a proper Lua exception
instead of just aborting.  To make that work without leaking memory and
file handles, the algorithm functions can now return NULL instead of the
new input pointer to indicate that they've aborted due to an error and left
an exception object on the Lua stack.  That gets turned into a Lua exception
only after appropriate cleanup has been done.