Remove old cruft from the data.filter code
Take out some stuff that was going to be an optimization to preallocate
just enough space for the output of the base64 algorithms when the size of
the input is known.  It was premature optimization.

Also take out a TODO for a '.util' table inside the main module table, which
was going to contain things like a 'sha1_transform' function.  I'll leave
that out until it's needed.