Add data.filter algorithms 'qp_encode' and 'qp_decode'
These do quoted-printable encoding, based on the specification in RFC 2045.
The implementation is mine from scratch, but the test suite is mostly
borrowed from the Perl module MIME::Base64 3.07, except that I've had to
adjust some of the expected results since my code behaves a bit differently.

The QP encoding should never produce lines longer than 76 bytes, not
including the end of line characters.  The Perl implementation will output
lines of 77 characters when the text was exactly 76 characters but there
was an additional '=' to indicate the lack of a newline in the input.  I
think that's a bug so I've avoided it, but I had to change the test suite
to reflect my opinion.

Currently encoding always canonicalises line ending characters (carriage
return and line feed) into a CRLF pair, which is probably the right thing
to do if you're producing text to go in an email.  You can supply an
alternative line ending, but I haven't tested that properly yet.

Decoding converts all line endings to Unix line feed only, which will
break binary data.  There should be some special mode for that.