Makefile target to build Win32 DLL for data.filter
I've added a new 'win32bin' target to the makefile which builds a DLL
called 'data.filter.dll'.  This seems to make the module work on Windows.
I've tested it on my Windows XP laptop, and in Wine under Debian unstable,
both using the LuaBinaries build of Lua 5.1.2.

I've never done anything like this before, and I'm a bit concerned that it
seemed to be simpler than a lot of the stuff on the web made out, so it's
possible I'm getting it completely wrong.  My theory though is that it's
relatively easy because the library I'm building doesn't need to be linked
against by anything by someone running a compiler (which seems to be far
more complicated under Windows than Linux), but only dynamically loaded
by the Lua executable.

The makefile is set up to build the DLL using the 'i586-mingw32msvc-cc'
cross compiler from the 'mingw32' package from Debian unstable, and to
look for the Lua library and headers in the current directory (so you need
to unpack the 'lua5_1_2_Win32_dll_lib.zip' file in the source directory).

I've also slightly refactored how the makefile adds debugging options, so
that the same ones can be applied to the Windows build.