Fix memory leaks when algorithm init function throws exception
I had a few algorithms where the 'algo_*_init' callback would throw an
error up with 'lua_error', and that would bypass cleanup code and leak
memory.  Turns out I wasn't testing that those errors were even thrown,
so I've added tests for all of them, and fixed the leak.

The init functions and a few other things now need to return a status
code so that I can detect the error in the places where the cleanup
needs to be done, and then pass it up to Lua.  This is the same scheme
as for the filtering callbacks.