Add built Unix library in '.libs' to Lua cpath
The crufty stuff in the makefile for automatically symlinking the shared
library to an appropriate name never really worked, so instead add the
expected name in '.libs' to 'pacakge.cpath' so that Lua can load it directly
from where libtool puts it, so that the test suite can load it.  Of course
once it's globally installed that won't be necessary for regular code, but
even then the local version will take precendence for the test suite.