Fix premature optimization bug; rename '_MODULE_NAME' to '_NAME'
I'm using '_NAME' now because it's the same as what the 'module' function
uses for the name of the module.

I've changed a some 'lua_pushlstring' calls to 'lua_pushstring' because
a couple of them had the wrong lengths, and they were only called once when
the module was required, so there's no need to micro-optimize them.
I've added tests that detected the buf.