datafilter: switch test suite from Lunit 0.3 to Lunit 0.4alpha
Also fixed a couple of accidental global variable accesses in
'test/options.lua' and 'test/pctenc.lua' where I'd forgotten to declare
an example value.

The diffs are quite big because I took the opportunity to move the test
data in some files to the top, so that I wouldn't need to separate the
declaration of the variable with giving it a value.  This seems neater.

All the tests still pass after these changes.