Remove the 'lunit' shell script and use 'lua' command directly to run tests
The shell script proved to be unnecessary, and was interfering with running
tests under valgrind and gdb.  I'm now just running 'lua' and feeding it the
one line of Lua code required to run the tests.  I can't use the '-e' option
for this because when it is run the command line arguments aren't available
from the special vararg expression (or the legacy 'arg' table).