Switch to Lua 5.3 and Lunitx
The original Lunit was never upgraded to work with Lua versions
newer than 5.1, but the Lunitx fork does, as well as having some
new features.

Reworked the tests to all be compiled into one file so that I can
run them more easily, without playing around echoing things into
the 'lua' command.  Simplifies the whole thing.  New filenames
for the individual test files to ensure they get put together in
the right order, partiularly the first and last files.

The last file, test/99_finish.lua, will run all the tests. This
is instead of the Lunitx provided way of using an atexit() function,
because that leaks memory when it does 'os.exit()'. That can't be
fixed by passing the second 'close' argument to it, because it's
running inside a GC finalizer, so attempting to have it close the
Lua state will seg fault.